Axe build system
I added the axe build system, that you can build a axe with 2 parts: The body and head. Each one can apply differents buffis that it shows on the machine UI!

+++ LumberjackGame/Assets/Scripts/SellItensSystem/Qol_BuySystem.cs
@@ -45,11 +45,11 @@ public class Qol_BuySystem : MonoBehaviour
         }
     }
 
-    public void CollectItem(Hec_Stats hecStats, InventoryItem item, int quant)
+    public void CollectItem(Hec_Stats hecStats, InventoryItem item, float quant)
     {
         hecStats.Hec_invent.hectorInventory.AddItem(quant, item);
     }
-    public void RemoveItem(Hec_Stats hecStats, InventoryItem item, int quant)
+    public void RemoveItem(Hec_Stats hecStats, InventoryItem item, float quant)
     {
         hecStats.Hec_invent.hectorInventory.RemoveItem(quant, item);
     }
